\name{NameNeedle-package}
\alias{NameNeedle-package}
\alias{NameNeedle}
\docType{package}
\title{
  Implements Needleman-Wunsch simple global alignment algorithm
}
\description{
  The Needleman-Wunsch simple gap algorithm was one of the first methods
  introduced for global alignment of biological sequences. The same algorithm
  can be used to match cell line names or sample names from two related data
  sets; we provide examples in the documentation, using data that accompanies
  this package.
  
  While the \pkg{NameNeedle} package can be used for biological sequence
  alignment, the \pkg{Biostrings} package from Bioconductor contains much more
  sophisticated tools for that problem.
}
\details{
  \tabular{ll}{
    Package: \tab NameNeedle\cr
    Type: \tab Package\cr
    Version: \tab 1.1.1\cr
    Date: \tab 2012-05-01\cr
    License: \tab GPL Version 2 or newer\cr
    LazyLoad: \tab yes\cr
  }
}
\references{
  Wang J, Byers LA, Yordy JS, Liu W, Shen L, Baggerly KA, Giri U, Myers
  JN, Ang KK, Story MD, Girard L, Minna JD, Heymach JV, Coombes KR.\cr
  \emph{Blasted cell line names.}\cr
  Cancer Inform. 2010; 9:251--5.
}
\author{
  Kevin R. Coombes \email{krc@silicovore.com},
  P. Roebuck \email{proebuck@mdanderson.org}
}
\keyword{package}

